requirements.txt
google-api-python-client
newsapi-python
pandas
vaderSentiment
sqlalchemy
streamlit
plotly